requirements.txt
# HearLoveen Complete Package - Python Requirements
# Install: pip install -r requirements.txt

# ===========================
# CORE DATA SCIENCE
# ===========================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# ===========================
# VISUALIZATION
# ===========================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ===========================
# STATISTICAL ANALYSIS
# ===========================
statsmodels>=0.14.0
lifelines>=0.27.0  # Survival analysis

# ===========================
# MACHINE LEARNING
# ===========================
torch>=2.0.0
tensorflow>=2.13.0
transformers>=4.30.0

# ===========================
# EXPLAINABLE AI
# ===========================
shap>=0.42.0
lime>=0.2.0

# ===========================
# API & WEB
# ===========================
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# ===========================
# DATABASE
# ===========================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
alembic>=1.11.0  # Database migrations

# ===========================
# UTILITIES
# ===========================
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.65.0

# ===========================
# TESTING
# ===========================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ===========================
# DATA VALIDATION
# ===========================
great-expectations>=0.17.0

# ===========================
# AUDIO PROCESSING
# ===========================
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
pydub>=0.25.0

# ===========================
# NLP & SPEECH
# ===========================
nltk>=3.8.0
spacy>=3.6.0
phonemizer>=3.2.0

# ===========================
# MONITORING & LOGGING
# ===========================
loguru>=0.7.0
sentry-sdk>=1.28.0

# ===========================
# CLOUD & DEPLOYMENT
# ===========================
azure-storage-blob>=12.17.0
boto3>=1.28.0  # AWS SDK
google-cloud-storage>=2.10.0

# ===========================
# BUSINESS INTELLIGENCE
# ===========================
openpyxl>=3.1.0  # Excel support
xlsxwriter>=3.1.0
python-pptx>=0.6.21
python-docx>=0.8.11

# ===========================
# VERSION INFO
# ===========================
# Python >= 3.9 required
# Tested on Python 3.10 and 3.11

# ===========================
# OPTIONAL (GPU Support)
# ===========================
# torch-tensorrt  # For TensorRT optimization
# onnx>=1.14.0  # For ONNX model export
# onnxruntime-gpu>=1.15.0
